[interp] Bring back inlining support
This time we enable inlining of methods with multiple basic blocks. We implement inlining by merging the cbb of the caller method with the start bb of the inlined method. Code generation continues from the unique exit bblock of the inlined method. We don't yet support inlining methods that have multiple return points.