[interp] Initialize stack state of all bblocks
Before the bblock addition, we were iterating instructions from one bblock to the next one, inheriting the stack state. We only needed to set it for targets of branches. Now we scan every bblock independently from the othersi for simplicity, so all bblocks need to have the stack state initialized (including ones following a conditional branch).