[interp] Add basic block support (#20328)
* [interp] Rename method to be more sugestive

* [interp] Add beginning of basic block support

Emit instructions inside basic blocks. When emitting the code, we iterate over basic blocks in code order to get the full list of instructions. All optimizations are currently disabled.

* [interp] Remove is_bb_start

It had weird semantics and was unreliable. For two consecutive IL instructions, we can check the offset_to_bb array to see if the second instruction is associated with a different basic block. For IR instructions, there is no need to check that they are in the same basic block, since instructions are linked together only if they belong to the same basic block.

* [interp] Re-enable basic block optimizations

* [interp] Resolve relocations based on basic block native offsets

Rather than the mapping from IL offset to native offset, which wouldn't work for inlined methods.

* [interp] Move stack state tracking information to basic blocks

For every bbstart IL offset we were saving the stack state information, when entering the basic block. Move this information over to the actual basic block where it belongs. Aside from the simplification, the previous code wouldn't work when running optimizations on basic blocks of inlined methods, since all il offset mapping is already lost.

* [interp] Add debug print for bb

* [interp] Remove unnecessary code

We no longer need to have il offsets set for instructions since we resolve relocations directly based on basic blocks. Even empty basic blocks would have a native offset associated with them when emitting compacted instructions.

* [interp] Don't access clause_indexes when not inlining

It is invalid for the inlined method. We don't inline methods with clauses

* [interp] Check for null before accessing next instruction

* [interp] Bring back inlining support

This time we enable inlining of methods with multiple basic blocks. We implement inlining by merging the cbb of the caller method with the start bb of the inlined method. Code generation continues from the unique exit bblock of the inlined method. We don't yet support inlining methods that have multiple return points.

* [interp] Add inlining depth limit

* [interp] Simplify interp_insert_ins

* [interp] Simplify computation of clause ranges

The per instruction IL offset tracking was complicated and prone to bugs. We use it only for debug features now.

* [interp] Fix debugger tests

This commit does the following:
- it removes the wrongly populated preds array. We instead allocate the set of basic blocks in get_basic_blocks and then we link the basic blocks as we generate the code, the same way as JIT. We fully link bblocks with in/out, which will also be useful for future optimizations.
- we generate sdb intr locations the same way as JIT. This brings us closer to having consistency regarding the interruption point locations.

* [mini] Add more detail to OP_SEQ_POINT description

Helps to compare single step seq points between jit and interp

* [interp] Extract initialization of EH clause bblocks

get_basic_blocks is meant only to create the list of IL bblocks.

* [interp] Disable interp whitebox tests

Generated interpreter code and structure will change a lot so it doesn't make sense to maintain these tests with each change

* [interp] Only consider for tailcall optimization if prefix is present

This behaviour is the same as JIT. Tailcall optimization with interpreter is probably unreliable.

* [interp] Initialize stack state of all bblocks

Before the bblock addition, we were iterating instructions from one bblock to the next one, inheriting the stack state. We only needed to set it for targets of branches. Now we scan every bblock independently from the othersi for simplicity, so all bblocks need to have the stack state initialized (including ones following a conditional branch).