[interp] Fix offset_to_bb pointing to dead BBs
Dead BBs won't have their native_offset set, because they are no longer iterated over when emitting final code. This native_offset is needed when setting the clause ranges in the jit info. When we kill a bblock, make sure we replace all references in the offset_to_bb array to the following bblock.