[mono] Enable startup stats timer on wasm
I recommend reviewing each commit individually with the corresponding commit message. If desired I can split this into two PRs.

The end result is that this adds a new timer to the startup stats that works by getting the time early in Mono startup and then comparing with that at the time of stat dumps, as opposed to reading in process information. This means it should be easy to get working on all desired platforms.

Contributes to https://github.com/mono/mono/issues/18754