[interp] Remove dead basic blocks and merge adjacent ones (#20354)
* [interp] Dump bb index during verbose compilation

* [interp] Fix stack height of first bblock

* [interp] Add missing bblock link for switch case

* [interp] Create new bblock after rethrow

* [interp] Fix iteration over instructions during deadce

* [interp] Remove unused arg to interp_clear_ins

* [interp] Optimize away unnecessary bblocks

Merge consecutive bblocks and remove unreachable ones.

Don't optimize away blocks at the start of EH clauses.

* [interp] Fix offset_to_bb pointing to dead BBs

Dead BBs won't have their native_offset set, because they are no longer iterated over when emitting final code. This native_offset is needed when setting the clause ranges in the jit info. When we kill a bblock, make sure we replace all references in the offset_to_bb array to the following bblock.

* [interp] Make dump_interp_inst add newline by default

Calling it in the debugger does nothing due to lacking newline terminator.

* [interp] Fix instruction inserting during optimiztion passes

interp_insert_ins inserts into current basic block, which is only valid during code gen. We need to explicitly pass the actual bb that we are currently iterating.

* [interp] Unlink bblocks during cond br lowering

We run the bblock optimization pass during cprop iteration. If we are able to merge two basic blocks, then we redo cprop, since we could discover more optimizations.

* [interp] Log also total executed instructions

* [interp] Add support for inlining methods with multiple return points

As with jit, we create an exit bb (only during inlining) and have all return instruction branch to it instead.

* [interp] Remove unnecessary code

We always clear the stack when we start optimizing a new basic block.

* [interp] Ignore conditional branches to next ip during codegen

It is redundant and we were linking the same bblocks twice (once for target branch and once for next bblock), which is not allowed.

* [interp] Fix leak of bblock stack state

* [interp] Fix access of native offset of IL in dead bblock at method end

* Disable tests which to a lot of allocations

And can lead to OOM