Remove empty transform file.
This file was added in dotnet/runtime#41913. It causes git to refuse to check out master on Windows without disabling the `core.protectNTFS` config and when checked out on an NTFS drive, is automatically renamed to not have a period at the end, resulting in a change that renames the file to `transform` always being present in the working directory.

These both cause headaches with the developer inner loop.