[metadata] Assorted enhancements for some CIL image formats (#20405)
* [metadata] Try to set MonoImage guid from modules table for metadata_only images

If the image has a modules table, and a guid heap, set MonoImage:guid when
loading the image, even if it is metadata_only.  In minimal delta images this is
true.

* [metadata] Minimal delta files have 4 byte indices, always

* [metadata] Don't populate MonoImage:module_name for minimal deltas

* [metadata] Always store precise string heap size in MonoImage:heap_strings

In minimal delta images, the String heap is appended to the heap of the
base image.  But in the physical base image, the String heap is zero-padded to
a size that's a multiple of 4 bytes.

Compute the precise String heap size at base image load time.