Support loading assemblies and symbols into debugger via callback (#20396)
This PR adds support for loading assemblies/symbols into the debugging via an `assembly_loaded` callback as opposed to the current model which loads assemblies/symbols when the application is initialized via the `loaded_files` singleton.

Changes in this PR include:
- Splitting up debugger-test and lazy-debugger-test projects to support testing
- Adding `mono_wasm_add_files` method to signal debugger from client using console.log
- Added assembly_load callback to WASM debugger agent
- Add handling for `ADD_ASSEMBLY_PDB` event in MonoProxy

Validated via unit tests and StandaloneApp in Blazor. LazyAssemblyLoader needs to be modified to support this feature. I'll submit a PR to the aspnetcore repo for that.

Co-authored-by: captainsafia <captainsafia@users.noreply.github.com>