Tests fixup
Delete test that invokes the runtime in a loop forever.
This is just exercising a race between mono_thread_attach and the
runtime shutdown.  Instead update the invoke_foreign_thread test to
loop a few times to check that attach/detach loops are ok.

In the deadlock test, wait for the foreign thread to finish calling
the runtime before the test thread returns from native back to managed
to avoid a race between shutdown and the invoke.