[interp] Fix vararg calls
Simplify the code on the call site, arguments are passed identically as in a normal call. However, a vararg method starts with a MINT_INIT_ARGLIST, which saves the varargs for later use. These args are stored in a localloc section, so the frame structure of a vararg method is still statically determined at compile time.