[interp] Unify execution and valuetype stacks  (#20500)
* [interp] Simplify constrained call checks

ECMA states 3 possible cases : constrained class is a reference, constrained class is a value type and it implements method to be called, constrained class is a value type and it doesn't implement method to be called.

* [remoting] Fix invalid IL in xdomain-dispatch wrapper

The target of the brfalse branch was reached with a stack height of 1 from the conditional branch instruction and with a stack height of 0 from the follow through block.

* [interp] Unify execution and valuetype stacks

Before this change, an InterpFrame contained 3 regions of data : args + locals, execution stack, valuetype stack. Each entry on the execution stack was a stackval structure. The space for valuetypes, was allocated separately, since they have various sizes. When pushing a valuetype on the stack, we allocated first the space for it on the vtstack and then pushed the address of the region on the executino stack. This change merges the execution stack with the valuetype stack, meaning we push now variable sized data on the execution stack. In order to keep track of the current stack location, whenever we push a type on stack, during transform phase, we also keep track of the offset where this value will reside on stack, as well as the size it occupies. All callsites need to be informed how much they need to pop the stack for the arguments. While called code can access this space normally (the args are special locals belonging to the frame and are accessed directly as such), external code needs a new mechanism to detect each argument for a given frame. This is achieved with the lazily initialized arg_offsets array on an InterpMethod. The method doesn't need to be compiled for this array to be correctly initialized.

Why :
- this simplifies handling of valuetypes, their storage follows the same rules as a normal objref/primitive type
- removes the common use of the vt_sp variable. The compiler no longer needs to reserve it in a register during the switch loop, we no longer need to save it with each call. The sp and ip become now the only variables describing the execution state in a method.
- the flow of the data on the execution stack is well behaved now (with the exception of a few opcodes that update directly based on the stack offset). We were using the vtstack for some magic storage (MINT_VTRESULT for example)
- this makes it such that the stack offset of every value is easily known at compile time, making it possible to completely drop the execution stack approach, and have every opcode have a unique dreg and a list of sregs (which are mapped to a certain stack offset). This will enable more advanced optimizations during compile stage.

* [interp] Fix vararg calls

Simplify the code on the call site, arguments are passed identically as in a normal call. However, a vararg method starts with a MINT_INIT_ARGLIST, which saves the varargs for later use. These args are stored in a localloc section, so the frame structure of a vararg method is still statically determined at compile time.

* [interp] Remove dead opcode