[interp] Switch all instructions to operate on local offsets
In order to facilitate future optimizations and also reduce the total number of instructions (since ldloc/stloc/mov are mostly redundant with this change). Instead of having instructions pop and push to the stack, they now receive explicit offsets for dreg and sregs. The offset is relative to frame->stack. Call args/ret are handled through a single dreg (this is a special CallArgs dreg). As with other locals, this dreg will be resolved to an offset. However, in addition to the optional return value of the call, at this offset, the call args will reside one after the other (This means the called frame will have the stack base at this offset, and the arguments are directly part of its local space). A local that holds an arg value to a call will never be referenced again (as a sreg in another instruction), but its offset will be in the param area of the call, and we will never optimize it out, since the call keeps it alive.

Since instructions only operate on stackvals, all IL locals of primitive types now occupy sizeof (stackval), instead of their real size. This enables them to be used directly as the source / destination of an instruction. Currently, there are two types of locals : normal locals and execution stack locals. While normal locals occupy space on the stack on the first come first served basis, execution stack locals have a predefined stack offset (relative to the start of the execution stack that we used to have before, which is computed during the initial code generation, mapping directly to the IL stack). Their real offset is resolved only after all the normal locals have their offset determined. Execution stack locals will become normal locals in the future, once we have a local offset allocator that takes into account the liveness region of locals.