[interp] Re-enable copy propagation
Replace the previous cumbersome cprop with a new simpler and more generic pass. It iterates on the instructions of each basic block, one at a time, saving the definition for every destination local. If a local is used as a sreg and its definition is a MOV, then we can forward through the copy. The deadce pass will remove locals that are not referenced anymore.

We add INTERP_LOCAL_FLAG_CALL_ARGS flag to locals that represent call args, since deadce would normally remove them. We generate moves when transitioning from a basic block to another while having a stack state in fixup_newbb_stack_locals. This is serves the same purpose as a phi instruction, in the future we should probably change to SSA and further simplify parts of the implementation. The fact that we don't have our own local offset allocator for execution stack locals, means that we are not free to extend liveness of these type of locals, overcomplicating the cprop pass. For now.