Add support for external icalls not using a wrapper.
Add support to register an external icall getting minimal runtime
overhead when called. An icall that can be used without a wrapper
needs to follow specific rules in order to correctly collaborate
with runtime. All functions are currently internal only and only
accessible by runtime or embedders using static linking and fully
acknowledge that this is not an optimization available as a public
API.