Add support for global aggressive-linining optimization. (#20641)
* Add support for global aggressive-linining optimization.

Currently we do support aggressive inlining per method using attributes.
This commit adds a new global mono optimization flag that does similar
thing, but whiout the need to annotate methods. There is still one big
difference and that is the effect of cost. The new global optimization
does not override cost evaluation, so if a method is to expensive to
inline with regards to inline limit, it won't be inlined. In order to
get the full "forced" inlining. method attributes should still be used.

This new optimization is not default and needs to be enabled using:

--optimize=aggressive-inlining

* Keep default inline limit on none netcore builds.