Fix bad case-insensitive ASCII equality check
Fixes https://github.com/dotnet/runtime/issues/45889. There was an edge case where the characters ``'`'`` and `'@'` would inadvertently compare as equal under _OrdinalIgnoreCase_. This PR also introduces a unit test to run through all combinations of ASCII chars to look for other edge cases, and it looks like this was the only edge case.

My AMD-based dev box is reporting a tiny perf improvement: maybe 1 cycle or so. I'm not able to measure this with much accuracy, so I'd depend on the perf lab to let us know if something has gone horribly wrong. Since we're reducing the overall code size I wouldn't expect throughput to regress.

This will need to be backported to 5.x and 3.x.