[Mono] Add an attribute to layout runtime visible classes
## `MonoRuntimeLayoutAware` attribute

Add a new attribute `MonoRuntimeLayoutAware` to be used on classes that are exposed to the runtime.  The fields of the class that are visible to the runtime go into a separate struct, and the attribute ensures that the runtime places the field first in the class.

Usage:

```csharp
[MonoRuntimeLayoutAware(nameof(_runtime)]
sealed class MyClass {
   [StructLayout (LayoutKind.Sequential)]
   private struct RuntimeLayout {
     IntPtr ptr;
     IntPtr ptr2;
  }

  private RuntimeLayout _runtime;     // runtime will put this first in MyClass
  private object _some_other_field;   // other fields can be trimmed by the linker, etc
}
```

Caveats:

* this only works for classes the derive from `object` or from classes that don't add any instance fields.
* this won't work for subclasses that also need to be visible (e.g. most of `System.Reflection.Emit`, and unfortunately `System.ArgumentException`)

## `Mono.ICallArg<T>` ref struct

Currently the IL Linker is conservative and it won't remove any fields from a class that is passed to an `InternalCall`. (This is reasonable behavior.  Although we could potentially teach the linker that for a `MonoRuntimeLayoutAware` class it's ok to just keep the `MonoRuntimeLayoutAware` field, and consider the rest as trimmable)

As a workaround, but also because it aligns us a bit closer with the CoreCLR [`QCallHandles`](https://github.com/dotnet/runtime/blob/master/src/libraries/System.Private.CoreLib/src/System/Runtime/CompilerServices/QCallHandles.cs#L24) add a new `ref struct ICallArgs<T>` type.

In places where we would pass, `Thread` to an internall call, now pass `Mono.ICallArg<Thread>` which is, effectively just `MonoInternalThread *volatile *` on the native side - and thus compatible with all our existing `HANDLES` (and some `NOHANDLES`) icalls.

This is enough to get the linker to consider the fields of `Thread` as trimmable.  Currently all the trimmable fields of `Thread` are actually kept alive by other code in the BCL, but if we add `private IntPtr garbage` to `Thread`, the linker will remove it.  So the `ICallArg<Thread>` trick is working.