[sgen] Use new MonoClassKind MONO_CLASS_GC_FILLER in mono_class_create_array_fill_type
The array fill type `MonoClass` is used to populate the vtable of sgen nursery free space.

Previously we used a static `MonoClass` that was zero-initialized. To make it easier to spot pointers to freed nursery memory, initialize the `class_kind` to a new constant `MONO_CLASS_GC_FILLER`.

Check for the new value explicitly in `class-accessors.c`