[mono][debugger] Implementing mscordbi to support iCorDebug on mono runtime
There are some things hardcoded yet, because we still need to decide how we will launch the process and decide the debugger port.
Using a dbgshim changed to accept mono runtime as a coreclr runtime with this version is possible to:
- Watch simple variables
- Watch string variables
- Watch array variables
- Watch object variables
- Watch this
- Step over / step into / step out
- Stop / Continue
- Evaluate static methods (with and without parameters)

For this implementation we decided to use metadata reader from coreclr, this is great because I could reuse a lot of code from mscordbi of coreclr.
I'm also using the log infrastructure from coreclr, if we start vsdbg-ui and set these parameters on environment:
COMPlus_LogFile=c:\thays\example.txt
COMPlus_LogToFile=1
COMPlus_LogLevel=8
COMPlus_LogFacility=0x00000200
COMPlus_LogEnable=1

We will see the log message with the icordebug messages which are not implemented. This may help to continue the implementation.

TODO: delete objects, count references and delete when it gets on 0.