Marshal class arguments as object references by default.
The MSDN documentation states that classes are always marshalled as
interfaces. This is wrong. In fact, this only happens if the class is an
interface or if UnmanagedType.Interface is used, otherwise the class is an
object reference (which is basically a pointer to a struct in C).

The tests here should pretty much prove this behavior; and there are plenty
of examples in MSDN and elsewhere where a non-interface class is used to
marshal a struct pointer (such as the POINT struct from native Windows API).