Add support for Apple Silicon on Xamarin.Mac SDK (#20597)
This PR implements changes that lets us build Xamarin.iOS/Xamarin.Mac SDKs so that it is possible for a Xamarin.Mac project to target x86_64+arm64 (i.e. Apple Silicon). Most of the changes are related to the mac Sdk build, but there are some extra changes as well :

- Fix configure to not automatically target iOS when targeting aarch64
  - Necessary to compile a cross-compiler targetting macOS arm64
- Fix aot compiler to target macOS arm64 when running on x64 (missing arch parameters)
- Support offset tool for aarch64
  - Necessary to compile macOS arm64 cross-compiler
- Fix pthread_jit_write_protect_np and preadv usage
  - Missing define meant the function wasn't found (at least on Xcode 12.2)
  - Restrict usage to macOS
- Fix crash on launch on iOS/tvOS arm64 caused by invalidly setting `MAP_JIT` on those platforms even though it's not a valid option.
- Adapt macOS SDK
  - Build libraries for x64+arm64 and lipo them together
  - Create a aarch64 cross-compiler for macOS, similar to the iOS Sdk

The arm64 support will only work with Xcode 12.2. To support in Xamarin.Mac, it also doesn't work if built from stable branches (because the rest of the arm64 support hasn't made it to the latest stable branch from what I can figure out)

Co-authored-by: Alexander Köplinger <alex.koeplinger@outlook.com>