Removing the building of the sgen runtime for android.
The motivation for this change was due to the fact that the libmono-native library needs to link against the mono runtime because it needs symbols from it. When the sgen runtime is built it defaults to linking against sgen and therefore at runtime when the libmono-native library is used in a boehm conext it still looks for the sgen runtime to get access to the symbols it wants.