Allow shifting managed pointers.
This can happen with code that uses "transient pointers" as
specified by ECMA-335. These are supposed to be usable wherever
native ints (STACK_PTR) or managed pointers (STACK_MP) can be used.
Mono doesn't have a concept of transient pointers and tends to
treat them as STACK_MP, which does not allow shifting.

The whole thing is very confusing, but the end result is that we
see sensible code in winehq bug 50246 that does a subtraction
involving transient pointers, and Mono decides the result of the
subtraction is STACK_MP, because it happened to map one to
STACK_PTR and one to STACK_MP. It then fails when the STACK_MP
result is shifted. Rather than try to implement ECMA-335's
confusing instructions correctly, which would require substantial
changes and probably not even produce something closer to .NET
Framework, the safest solution is to just make things more
permissive.